+++ NINA.Sequencer/Trigger/Autofocus/AutofocusAfterHFRIncreaseTrigger.cs
@@ -116,6 +116,17 @@ namespace NINA.Sequencer.Trigger.Autofocus {
             }
         }
 
+        private bool trendPerFilter = true; // default true to keep the original behaviour creating an HFR trend per filter 
+
+        [JsonProperty]
+        public bool TrendPerFilter {
+            get => trendPerFilter;
+            set {
+                trendPerFilter = value;
+                RaisePropertyChanged();
+            }
+        }
+
         private double originalHFR;
 
         public double OriginalHFR {
@@ -177,7 +188,7 @@ namespace NINA.Sequencer.Trigger.Autofocus {
                 imageHistory = imageHistory.Where(point => point.Id > lastAF.Id).ToList();
             }
 
-            if (fwInfo != null && fwInfo.Connected && fwInfo.SelectedFilter != null) {
+            if (fwInfo != null && fwInfo.Connected && fwInfo.SelectedFilter != null && TrendPerFilter == true) {
                 //Further filter the history to only considere items by the current filter
                 Filter = fwInfo.SelectedFilter.Name;
 
@@ -240,7 +251,7 @@ namespace NINA.Sequencer.Trigger.Autofocus {
         }
 
         public override string ToString() {
-            return $"Trigger: {nameof(AutofocusAfterHFRIncreaseTrigger)}, Amount: {Amount}";
+            return $"Trigger: {nameof(AutofocusAfterHFRIncreaseTrigger)}, Amount: {Amount}, TrendPerFilter: {TrendPerFilter}";
         }
 
         public bool Validate() {