+++ RELEASE_NOTES.md
@@ -82,6 +82,8 @@ More details at <a href="https://nighttime-imaging.eu/donate/" target="_blank">n
     - Fix issue where saving sequences after multi-file load overwrote only one XML
 - **Time source failure reasons**
     - When a time source fails to provide a value - e.g. that there is no astronomical twilight today - the validation will also include the reason for the failure. This will help to understand why a time source is not providing a value.
+- **Autofocus after HFR Increase Trigger**
+    - new Trend per Filter checkbox to consider HFR Trend per filter (default) or across all filters
 
 ### **Guiding & Tracking**
 - **Looping Mode Behavior**